requirements.txt
acme==1.1.0
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2019.11.28
chardet==3.0.4
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography>=2.8,<3.4
pyjwt[crypto]>=2.5.0,<3.0
distlib==0.3.4
distro==1.4.0
filelock==3.7.1
future==0.18.2
idna==2.8
josepy==1.2.0
mock==3.0.5
parsedatetime==2.4
pbr==5.4.5
platformdirs==2.5.2
ply==3.11
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
six==1.14.0
ssh-import-id==5.10
urllib3==1.25.8
virtualenv==20.15.1
zope.component==4.3.0
zope.event==4.4
firebase-admin==6.2.0
django-cors-headers==4.2.0
fcm-django==2.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2

Django~=4.2.3

djangorestframework~=3.14.0
asgiref~=3.7.2
channels~=4.0.0
ffmpeg~=1.4
Pillow~=10.0.0
moviepy~=1.0.3
phonenumbers~=8.13.17

opencv-python~=4.8.0.74
pilkit~=2.0